Empty input now emits an event
Because empty input didn't count as an event before, an event would be emitted for the first character as the user holds down the back button to delete all of the text. With this change, an event will be emitted for empty input; however this event will map to Observable.never().